+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
   # integrations
   "integrations/actix",
   "integrations/axum",
+  "integrations/utils",
 
   # libraries
   "meta",
@@ -19,17 +20,18 @@ members = [
 exclude = ["benchmarks", "examples"]
 
 [workspace.package]
-version = "0.1.3"
+version = "0.2.0-alpha"
 
 [workspace.dependencies]
-leptos = { path = "./leptos", default-features = false, version = "0.1.3" }
-leptos_dom = { path = "./leptos_dom", default-features = false, version = "0.1.3" }
-leptos_macro = { path = "./leptos_macro", default-features = false, version = "0.1.3" }
-leptos_reactive = { path = "./leptos_reactive", default-features = false, version = "0.1.3" }
-leptos_server = { path = "./leptos_server", default-features = false, version = "0.1.3" }
-leptos_config = { path = "./leptos_config", default-features = false, version = "0.1.3" }
-leptos_router = { path = "./router", version = "0.1.3" }
-leptos_meta = { path = "./meta", default-feature = false, version = "0.1.3" }
+leptos = { path = "./leptos", default-features = false, version = "0.2.0-alpha" }
+leptos_dom = { path = "./leptos_dom", default-features = false, version = "0.2.0-alpha" }
+leptos_macro = { path = "./leptos_macro", default-features = false, version = "0.2.0-alpha" }
+leptos_reactive = { path = "./leptos_reactive", default-features = false, version = "0.2.0-alpha" }
+leptos_server = { path = "./leptos_server", default-features = false, version = "0.2.0-alpha" }
+leptos_config = { path = "./leptos_config", default-features = false, version = "0.2.0-alpha" }
+leptos_router = { path = "./router", version = "0.2.0-alpha" }
+leptos_meta = { path = "./meta", default-feature = false, version = "0.2.0-alpha" }
+leptos_integration_utils = { path = "./integrations/utils", version = "0.2.0-alpha" }
 
 [profile.release]
 codegen-units = 1
