+++ meta/src/lib.rs
@@ -278,6 +278,22 @@ impl MetaContext {
     }
 }
 
+/// Extracts the metadata that should be used to close the `<head>` tag
+/// and open the `<body>` tag. This is a helper function used in implementing
+/// server-side HTML rendering across crates.
+pub fn generate_head_metadata(cx: Scope) -> String {
+    let meta = use_context::<MetaContext>(cx);
+    let head = meta
+        .as_ref()
+        .map(|meta| meta.dehydrate())
+        .unwrap_or_default();
+    let body_meta = meta
+        .as_ref()
+        .and_then(|meta| meta.body.as_string())
+        .unwrap_or_default();
+    format!("{head}</head><body{body_meta}>")
+}
+
 /// Describes a value that is either a static or a reactive string, i.e.,
 /// a [String], a [&str], or a reactive `Fn() -> String`.
 #[derive(Clone)]
