+++ router/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "leptos_router"
-version = "0.1.3"
+version = "0.2.0-alpha"
 edition = "2021"
 authors = ["Greg Johnston"]
 license = "MIT"
